Remove all deprecated code to the same file
