+++ utils/graphile-config/src/index.ts
@@ -5,8 +5,14 @@ import { sortWithBeforeAfterProvides } from "./sort.js";
 export { GraphileConfig };
 
 export { FunctionalityObject, orderedApply } from "./functionality.js";
-export { applyHooks, AsyncHooks, HookObject } from "./hooks.js";
-export type { PluginHook, PluginHookObject } from "./interfaces.js";
+export {
+  applyHooks,
+  AsyncHooks,
+  HookObject,
+  PluginHook,
+  PluginHookObject,
+} from "./hooks.js";
+export type { CallbackDescriptor, OrderedCallback } from "./interfaces.js";
 export type { MiddlewareNext, MiddlewareObject } from "./middleware.js";
 export { Middlewares } from "./middleware.js";
 export { isResolvedPreset, resolvePresets } from "./resolvePresets.js";
